test: Pull methods tests out into their own files. Easier to manage before/after setup and teardown
